(리팩토링) 호감개수 Long 타입에서 int로 수정

+++ src/main/java/com/ll/gramgram/boundedContext/likeablePerson/repository/LikeablePersonRepository.java
@@ -11,5 +11,5 @@ public interface LikeablePersonRepository extends JpaRepository<LikeablePerson, 
 
     LikeablePerson findByFromInstaMemberIdAndToInstaMember_username(Long fromInstaMemberId, String toInstaMemberUsername);
 
-    Long countByFromInstaMemberId(Long id);
+    int countByFromInstaMemberId(Long id);
 }
